BUG: Fix incorrect tm_matrix in call to visitor_text #2060
Supply the old tm_matrix when flushing out `text` to the `visitor_text`
in `crlf_space_check`. The new one might already be changed and
unrelated to the current text.

Also add a test for the tm_matrix and cm_matrix that are given to
`visitor_text` when extracting text.
The test computes the coordinates of three letters in different
parts of a test page based on the matrices and checks, if they are
roughly where they should be.

Fixes #2059